Fixed happiness and added time delay to resource generators
Happiness now resets every time the game starts. While I can't finish it now, a rough version of a delay has been implemented on the dirt mine. This delay will be greater if the haven is unhappy, and speed up the delay when happiness is greater. Further testing is required to ensure successful implementation.