Fixed happiness
Happiness and productivity now work as intended. If an amenity makes a tile happy, it correctly affect the haven's happiness and productivity. However, this does not account for if a tile is no longer affected by an amenity, and therefore should not be happy. As such, the HappinessTracker script will need to be tweaked in order to account for the tile no longer being in the vicinity of an amenity.